+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @formwright/schema
 
+## 0.2.2
+
+## 0.2.1
+
 ## 0.2.0
 
 ### Minor Changes